refactor(core): extract tune() orchestration into writer-exempt ModelTuningMixin (H-0091)
Move the ~430-line tuning orchestration (tune + _validate_tune_inputs /
_resolve_search_space / _maybe_expand_boundary / _build_tune_objective /
_run_tune_round / _assemble_tuning_result / _get_tuning_state) from
core/model.py to a new core/_model_tuning.py (ModelTuningMixin), bringing the
facade from 1257 to 789 lines. Resolves the H-0077 conflict deferred from
#209: the read-only-mixin invariant covers only the diagnostic mixins
(plots/tables/persistence, enforced by _MIXIN_FILES); ModelTuningMixin is a
writer that runs during the mutating tune() lifecycle and is deliberately
outside that guard. _DEFAULT_METRICS moved to the shared _model_metrics.py.
Pure refactor — no public-API / FitResult / format_version change; full
tuning + retune suite unchanged.

Refs #237
