+++ CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Internal
 
+- **`tune()` orchestration extracted into a writer-exempt mixin** (H-0091, [#237](https://github.com/nbx-liz/LizyML/issues/237)). The ~430-line tuning orchestration (`tune` + `_validate_tune_inputs` / `_resolve_search_space` / `_maybe_expand_boundary` / `_build_tune_objective` / `_run_tune_round` / `_assemble_tuning_result` / `_get_tuning_state`) moved from `core/model.py` to a new `core/_model_tuning.py` (`ModelTuningMixin`), bringing the facade from 1257 to **789 lines** (< the 800 guidance). Resolves the H-0077 conflict deferred from #209: the read-only-mixin invariant applies only to the *diagnostic* mixins (plots/tables/persistence, enforced by the `_MIXIN_FILES` guard); `ModelTuningMixin` is a *writer* that runs during the mutating `tune()` lifecycle and is deliberately outside that guard. `_DEFAULT_METRICS` moved to the shared `core/_model_metrics.py`. Pure refactor — no public-API / `FitResult` / `format_version` change; the full tuning + retune suite is unchanged.
 - **Fail-closed leakage tests** (H-0085, [#207](https://github.com/nbx-liz/LizyML/issues/207)) — replaced a tautological OOF leakage assertion with a trap that records each fold estimator's training rows and asserts disjointness from its validation rows; added traps for per-fold calibration boundaries (calibrator fit on the train slice, not the scored valid rows) and inner-index containment (relative to each outer train fold).
 - **Plot optional-dependency guard de-duplicated** ([#218](https://github.com/nbx-liz/LizyML/issues/218)). The per-module `_require_plotly` / `_check_plotly` guards (two names, three message strings) are replaced by a single `lizyml/plots/_deps.py::require_plotly(...)`, now applied across **all** plot modules (`learning_curve` / `oof_distribution` / `classification` / `residuals` previously kept an inline `if _plotly is None` guard). Same `OPTIONAL_DEP_MISSING` behavior. The near-duplicate `_collect_is_data` (classification) / `_build_is_data` (residuals) IF-data collectors are hoisted into a single `lizyml/plots/_helpers.py::collect_is_data(...)`.
 - **`fit()` after `tune()` now logs an optimistic-bias warning** ([#218](https://github.com/nbx-liz/LizyML/issues/218)). A `fit()` following `tune()` reuses the identical deterministic CV splits used to select the params, so its reported OOF metrics are optimistically biased (documented user-side policy, BLUEPRINT §11.6, previously unenforced at runtime). The facade now emits one `event='fit.post_tune'` warning per such fit. Log-only; no contract change.
